+++ src/watchers/config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  TEMP_DIR_PATH: `${__dirname}/tmp`,
+  DOWNLOADED_DIR_PATH: `${__dirname}/downloaded`,
+};
